remove type cast

+++ implementations/other/RandomSeed.h
@@ -5,6 +5,6 @@
  * Time: O(1)
  */
 
-mt19937 rng((uint32_t) chrono::steady_clock::now().time_since_epoch().count());
+mt19937 rng(chrono::steady_clock::now().time_since_epoch().count());
 
 // uniform_int_distribution<int>(a, b)(rng) generates a random integer between a and b inclusive
